#46573 move endpoints for zaakInstanceLinks after request from Thomas
