+++ openzaak/src/main/kotlin/com/ritense/openzaak/security/config/ZaakInstanceLinkHttpSecurityConfigurer.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015-2022 Ritense BV, the Netherlands.
+ * Copyright 2015-2023 Ritense BV, the Netherlands.
  *
  * Licensed under EUPL, Version 1.2 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.ritense.objectenapi.security
+package com.ritense.openzaak.security.config
 
 import com.ritense.valtimo.contract.authentication.AuthoritiesConstants.USER
 import com.ritense.valtimo.contract.security.config.HttpConfigurerConfigurationException
@@ -22,15 +22,14 @@ import com.ritense.valtimo.contract.security.config.HttpSecurityConfigurer
 import org.springframework.http.HttpMethod.GET
 import org.springframework.security.config.annotation.web.builders.HttpSecurity
 
-class ObjectenApiHttpSecurityConfigurer: HttpSecurityConfigurer {
+class ZaakInstanceLinkHttpSecurityConfigurer : HttpSecurityConfigurer {
 
     override fun configure(http: HttpSecurity) {
         try {
             http.authorizeRequests()
-                .antMatchers(GET, "/api/v1/document/{documentId}/zaak/objecttype").hasAuthority(USER)
-                .antMatchers(GET, "/api/v1/document/{documentId}/zaak/object").hasAuthority(USER)
-                .antMatchers(GET, "/api/v1/document/{documentId}/zaak/object/form").hasAuthority(USER)
-        } catch(e: Exception) {
+                .antMatchers(GET, "/api/v1/zaakinstancelink/zaak").hasAuthority(USER)
+                .antMatchers(GET, "/api/v1/zaakinstancelink/document").hasAuthority(USER)
+        } catch (e: Exception) {
             throw HttpConfigurerConfigurationException(e)
         }
     }
