+++ zgw/objecten-api/src/test/kotlin/com/ritense/objectenapi/web/rest/ZaakObjectResourceTest.kt
@@ -16,28 +16,22 @@
 
 package com.ritense.objectenapi.web.rest
 
-import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
-import org.mockito.kotlin.mock
-import org.mockito.kotlin.whenever
 import com.ritense.form.domain.FormIoFormDefinition
 import com.ritense.objectenapi.client.ObjectRecord
 import com.ritense.objectenapi.client.ObjectWrapper
 import com.ritense.objectenapi.service.ZaakObjectService
-import com.ritense.objectenapi.web.rest.result.ZaakInstanceLinkDTO
 import com.ritense.objecttypenapi.client.Objecttype
-import com.ritense.openzaak.domain.mapping.impl.ZaakInstanceLink
-import com.ritense.openzaak.domain.mapping.impl.ZaakInstanceLinkId
-import com.ritense.openzaak.service.ZaakInstanceLinkService
 import com.ritense.valtimo.contract.json.Mapper
 import org.hamcrest.Matchers
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import org.mockito.kotlin.mock
+import org.mockito.kotlin.whenever
 import org.springframework.http.MediaType
 import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter
 import org.springframework.test.web.servlet.MockMvc
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get
 import org.springframework.test.web.servlet.result.MockMvcResultHandlers
-import org.springframework.test.web.servlet.result.MockMvcResultMatchers
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
 import org.springframework.test.web.servlet.setup.MockMvcBuilders
@@ -51,13 +45,11 @@ internal class ZaakObjectResourceTest {
     lateinit var mockMvc: MockMvc
     lateinit var zaakObjectService: ZaakObjectService
     lateinit var zaakObjectResource: ZaakObjectResource
-    lateinit var zaakInstanceLinkService: ZaakInstanceLinkService
 
     @BeforeEach
     fun init() {
         zaakObjectService = mock()
-        zaakInstanceLinkService = mock()
-        zaakObjectResource = ZaakObjectResource(zaakObjectService, zaakInstanceLinkService)
+        zaakObjectResource = ZaakObjectResource(zaakObjectService)
 
         mockMvc = MockMvcBuilders
             .standaloneSetup(zaakObjectResource)
@@ -195,90 +187,4 @@ internal class ZaakObjectResourceTest {
         return converter
     }
 
-    @Test
-    fun `should return ZaakInstaceLink for given zaakInstanceUrl`() {
-        // given:
-        val url = URI("http://localhost:8001/zaken/api/v1/zaken/df64a38e-566a-409d-8275-9207f70f79e7")
-        val documentId = UUID.randomUUID()
-
-        val entity = ZaakInstanceLink(
-            zaakInstanceLinkId = ZaakInstanceLinkId(
-                id = UUID.randomUUID()
-            ),
-            zaakInstanceUrl = url,
-            zaakInstanceId = UUID.randomUUID(),
-            documentId = documentId,
-            zaakTypeUrl = URI("http://example.com"),
-        )
-
-        //when:
-        whenever(zaakInstanceLinkService.getByZaakInstanceUrl(url)).thenReturn(
-            entity
-        )
-
-        //then:
-        val expectedDto = ZaakInstanceLinkDTO(
-            zaakInstanceUrl = url,
-            documentId = documentId
-        )
-
-        mockMvc.perform(
-            get("/api/v1/zaakinstancelink/zaak?zaakInstanceUrl=$url")
-                .characterEncoding(StandardCharsets.UTF_8.name())
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-        )
-            .andDo(MockMvcResultHandlers.print())
-            .andExpect(status().is2xxSuccessful)
-            .andExpect(jsonPath("$").isNotEmpty)
-            .andExpect(
-                MockMvcResultMatchers.content().json(
-                    jacksonObjectMapper().writeValueAsString(expectedDto)
-                )
-            )
-    }
-
-    @Test
-    fun `should return ZaakInstaceLink for given documentId`() {
-        // given:
-        val url = URI("http://example1.com")
-        val documentId = UUID.randomUUID()
-
-        val result = ZaakInstanceLink(
-            zaakInstanceLinkId = ZaakInstanceLinkId(
-                id = UUID.randomUUID()
-            ),
-            zaakInstanceUrl = URI("http://example1.com"),
-            zaakInstanceId = UUID.randomUUID(),
-            documentId = documentId,
-            zaakTypeUrl = URI("http://example2.com"),
-        )
-
-        //when:
-        whenever(zaakInstanceLinkService.getByDocumentId(documentId)).thenReturn(
-            result
-        )
-
-        //then:
-        val expectedDto = ZaakInstanceLinkDTO(
-            zaakInstanceUrl = url,
-            documentId = documentId
-        )
-
-        mockMvc.perform(
-            get("/api/v1/zaakinstancelink/document?documentId=$documentId")
-                .characterEncoding(StandardCharsets.UTF_8.name())
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-        )
-            .andDo(MockMvcResultHandlers.print())
-            .andExpect(status().is2xxSuccessful)
-            .andExpect(jsonPath("$").isNotEmpty)
-            .andExpect(
-                MockMvcResultMatchers.content().json(
-                    jacksonObjectMapper().writeValueAsString(expectedDto)
-                )
-            )
-
-    }
 }
